scripts/png2carray/requirements.txt
certifi==2025.8.3
cffi==2.0.0
cryptography==46.0.1
flake8==7.3.0
mccabe==0.7.0
pillow==12.1.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.2/libexec/wheel-0.45.1-py3-none-any.whl#sha256=b9235939e2096903717cb6bfc132267f8a7e46deb2ec3ef9c5e234ea301795d0
